% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_new_name.R
\name{assign_new_names}
\alias{assign_new_names}
\title{Assigning new names}
\usage{
assign_new_names(.df = "extdata/Soil_extraction.xlsx")
}
\arguments{
\item{.df}{path to a masshunter exportsheet in .xlsx or .xls format}
}
\description{
A function that imports a Masshunter Export sheet and assigns new colnames.
This mainly resolves the issue of joint cells in the Masshunter export sheet. (This function is not exported)
}
\details{
A function rather aimed at developers
}
\keyword{internal}
