% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_format_from_export.R
\name{create_new_format_from_export}
\alias{create_new_format_from_export}
\title{create_new_format_from_export}
\usage{
create_new_format_from_export(.df = "extdata/Soil_extraction.xlsx")
}
\arguments{
\item{.df}{path to a masshunter exportsheet in .xlsx or .xls format}
}
\description{
This function creates a gathered tibble with the complete meta-information of a measured sample new columns are: mass,mode,element,parameter,unit,analyte_type
}
\details{
The function returns an  tibble of a Masshunter Exportsheet (.xlsx imported with read_excel()). The created tibble can be used with all dplyr operations like filter, select, group_by ect. for furher data analyses.
the excelsheet should contain the calibration curve or at least the calibration blank measurement set to level1. For further processing, it is advised to wirte the samples name in the Sample name collum and comments in the comment section.
}
