% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metadata.R
\name{import_metadata}
\alias{import_metadata}
\title{import_metadata}
\usage{
import_metadata(path_batch = "man/readme/extdata/random_ME_run.b/")
}
\arguments{
\item{path_batch}{path to the a target batchfile}
}
\description{
the function imports the metadata of all measurements saved in a mashunter batchfile.
}
