% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mh_all_sample_csv.R
\name{import_mh_all_sample_csv}
\alias{import_mh_all_sample_csv}
\title{import_mh_all_sample_csv}
\usage{
import_mh_all_sample_csv(
  path_batch = "data/200207_MainInc_ME_1_a.b",
  modes_nr = 2
)
}
\arguments{
\item{path_batch}{the path to the target batch file.}

\item{modes_nr}{a numerical indicating the number of modes (e.g. NoGas, He, ect.) used in the batch}
}
\description{
import CPS and descriptive stats of all samples and modes from in the batch folder
}
