% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data_standardmaterials.R
\docType{data}
\name{standardmaterials}
\alias{standardmaterials}
\title{This is a dataframe containing reference values of certified reference materials list may be updated to the users needs.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 117 rows and 10 columns.
}
\usage{
standardmaterials
}
\description{
This is a dataframe containing reference values of certified reference materials list may be updated to the users needs.
}
\author{
Lorenz Gfeller \email{lokagf@pm.me}
}
\keyword{data}
