% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_col_names_pos.R
\name{which_col_names_pos}
\alias{which_col_names_pos}
\title{which_col_names_pos}
\usage{
which_col_names_pos(.path = "extdata/Soil_extraction.xlsx")
}
\arguments{
\item{.path}{path to a masshunter exportsheet in .xlsx or .xls format}
}
\description{
this function returns a vector containing column name positions (This function is not exported)
}
\details{
this is an essential function for create_new_function_export
}
