% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mh_sample_csv.R
\name{import_mh_sample_csv}
\alias{import_mh_sample_csv}
\title{import_mh_sample_csv}
\usage{
import_mh_sample_csv(
  csv_path = "extdata/200207_MainInc_ME_1_a.b/001SMPL.d/001SMPL_1.csv"
)
}
\arguments{
\item{csv_path}{path to the csv file of the sample}
}
\description{
import CPS and descriptive stats of a signle sample from in the batch folder
}
